feat(components): add club member info modal component
