+++ src/components/index.ts
@@ -4,3 +4,4 @@ export * from './main';
 export * from './modal';
 export * from './section';
 export * from './mangeUserBook';
+export * from './manageClub';
