+++ src/components/manageClub/index.ts
@@ -1,0 +1,1 @@
+export * from './ClubMemberInfoModal';
