+++ src/constant/url.ts
@@ -1,2 +1,3 @@
 export const CLUB = '/club';
 export const MANAGE_CLUB_BOOK = '/manage/club-book';
+export const MANAGE_CLUB = '/manage-club';
